+++ app/src/main/java/org/seemeet/seemeet/data/model/request/login/RequestKakaoLogin.kt
@@ -5,5 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class RequestKakaoLogin(
     @SerializedName("socialtoken")
     val socialToken: String,
-    val provider: String
+    val provider: String,
+    val fcm : String
 )
