+++ app/src/main/java/org/seemeet/seemeet/ui/registration/LoginMainActivity.kt
@@ -6,6 +6,9 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.ktx.Firebase
+import com.google.firebase.messaging.ktx.messaging
 import com.kakao.sdk.auth.model.OAuthToken
 import com.kakao.sdk.user.UserApiClient
 import kotlinx.coroutines.CoroutineScope
@@ -30,7 +33,12 @@ class LoginMainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
+
+        //fcm 토큰 발급.
+        getFireBaseInstanceId()
+
         initClickListener()
+
     }
 
     fun initClickListener() {
@@ -98,9 +106,11 @@ class LoginMainActivity : AppCompatActivity() {
                 val body = RetrofitBuilder.loginService.postKakaoLogin(
                     RequestKakaoLogin(
                         socialToken,
-                        "kakao"
+                        "kakao",
+                        SeeMeetSharedPreference.getUserFb()
                     )
                 )
+
                 SeeMeetSharedPreference.setToken(body.data.accessToken.accessToken)
                 // 이름,닉네임 입력 안했으면 입력화면으로 이동, 했으면 메인으로 이동
                 if (body.data.user.nickname.isNullOrEmpty()) {
@@ -120,6 +130,20 @@ class LoginMainActivity : AppCompatActivity() {
         SeeMeetSharedPreference.setUserId(list.nickname ?: return)
         SeeMeetSharedPreference.setLogin(true)
         SeeMeetSharedPreference.setUserName(list.username)
+    }
+
+    private fun getFireBaseInstanceId(){
+        Firebase.messaging.token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                Log.d("******firebaseToken_Main_FAILED", task.exception.toString())
+                return@OnCompleteListener
+            }
+
+            val token = task.result
+            Log.d("******firebaseToken_LoginMain_before", SeeMeetSharedPreference.getUserFb())
+            SeeMeetSharedPreference.setUserFb(token)
+            Log.d("******firebaseToken_LoginMain_now", SeeMeetSharedPreference.getUserFb())
+        })
     }
 
     companion object {
